refactor(FXA-2218): extract _parse_steps_for_json to core/steps.py
Mechanical refactor. _parse_steps_for_json moves from
commands/plan_cmd.py:129 to new src/fx_alfred/core/steps.py module so
that validate_cmd.py can consume it for D2/D3 checks (FXA-2218 Commit 2+)
without introducing a commands -> commands import.

Zero behaviour change. plan_cmd.py re-exports via 'from fx_alfred.core.steps
import _parse_steps_for_json' so existing internal call sites work unchanged.

- pytest: 665/665 pass
- ruff: clean
- pyright: 0 errors

Commit 1/9 of FXA-2218.
