+++ src/fx_alfred/core/steps.py
@@ -1,0 +1,29 @@
+"""Shared step-parsing helpers for SOP bodies.
+
+Extracted from commands/plan_cmd.py to avoid commands -> commands imports
+when validate_cmd.py needs the same parser (FXA-2218 CHG Commit 1).
+"""
+
+from __future__ import annotations
+
+import re
+
+from fx_alfred.core.phases import StepDict
+
+
+def _parse_steps_for_json(section_text: str) -> list[StepDict]:
+    """Extract steps as structured data for JSON output.
+
+    Returns list of {"index": int, "text": str, "gate": bool}.
+    Gate is true if step ends with "✓" or contains "[GATE]".
+    """
+    steps: list[StepDict] = []
+    for line in section_text.split("\n"):
+        stripped = line.strip()
+        m = re.match(r"^(?:###\s+)?(\d+)\.\s+(.+)", stripped)
+        if m:
+            index = int(m.group(1))
+            text = m.group(2)
+            gate = text.endswith("✓") or "[GATE]" in text
+            steps.append({"index": index, "text": text, "gate": gate})
+    return steps
